Split target metadata matrix tests
